merge 3.0-dev into 3.0

+++ SPECS/bind/bind.spec
@@ -9,11 +9,11 @@
 
 Summary:        Domain Name System software
 Name:           bind
-Version:        9.16.37
-Release:        1%{?dist}
+Version:        9.16.44
+Release:        2%{?dist}
 License:        ISC
 Vendor:         Microsoft Corporation
-Distribution:   Mariner
+Distribution:   Azure Linux
 Group:          Development/Tools
 URL:            https://www.isc.org/downloads/bind/
 Source0:        https://ftp.isc.org/isc/bind9/%{version}/%{name}-%{version}.tar.xz
@@ -33,7 +33,6 @@ Source14:       setup-named-softhsm.sh
 Source15:       named-chroot.files
 Patch9:         bind-9.14-config-pkcs11.patch
 Patch10:        bind-9.10-dist-native-pkcs11.patch
-Patch11:        CVE-2023-2828.patch
 
 BuildRequires:  gcc
 BuildRequires:  json-c-devel
@@ -229,13 +228,12 @@ Summary:        BIND utilities
 %prep
 %setup -q
 
-%patch9 -p1 -b .config-pkcs11
+%patch 9 -p1 -b .config-pkcs11
 cp -r bin/named{,-pkcs11}
 cp -r bin/dnssec{,-pkcs11}
 cp -r lib/dns{,-pkcs11}
 cp -r lib/ns{,-pkcs11}
-%patch10 -p1 -b .dist_pkcs11
-%patch11 -p1
+%patch 10 -p1 -b .dist_pkcs11
 
 libtoolize -c -f; aclocal -I libtool.m4 --force; autoconf -f
 
@@ -386,22 +384,26 @@ popd
 rm -f %{buildroot}%{_prefix}%{_sysconfdir}/bind.keys
 
 %pre
-if ! getent group named >/dev/null; then
-    groupadd -r named
-fi
-if ! getent passwd named >/dev/null; then
-    useradd -g named -d %{_sharedstatedir}/bind\
-        -s /bin/false -M -r named
+if [ $1 -eq 1 ]; then
+  if ! getent group named >/dev/null; then
+      groupadd -r named
+  fi
+  if ! getent passwd named >/dev/null; then
+      useradd -g named -d %{_sharedstatedir}/bind\
+          -s /bin/false -M -r named
+  fi
 fi
 
 %post -p /sbin/ldconfig
 %postun
 /sbin/ldconfig
-if getent passwd named >/dev/null; then
-    userdel named
-fi
-if getent group named >/dev/null; then
-    groupdel named
+if [ $1 -eq 0 ]; then
+  if getent passwd named >/dev/null; then
+      userdel named
+  fi
+  if getent group named >/dev/null; then
+      groupdel named
+  fi
 fi
 
 # Fix permissions on existing device files on upgrade
@@ -416,7 +418,7 @@ if [ $1 -gt 1 ]; then \
   done \
 fi
 Vendor:         Microsoft Corporation
-Distribution:   Mariner
+Distribution:   Azure Linux
 %ldconfig_scriptlets libs
 %ldconfig_scriptlets pkcs11-libs
 
@@ -615,6 +617,15 @@ fi;
 %{_mandir}/man8/named-nzd2nzf.8*
 
 %changelog
+* Thu Dec 14 2023 Neha Agarwal <nehaagarwal@microsoft.com> - 9.16.44-2
+- Fix resetting of passwd and group on package update
+
+* Wed Sep 27 2023 CBL-Mariner Servicing Account <cblmargh@microsoft.com> - 9.16.44-1
+- Auto-upgrade to 9.16.44 - Fix CVE-2023-3341
+
+* Wed Sep 20 2023 Jon Slobodzian <joslobo@microsoft.com> - 9.16.37-2
+- Recompile with stack-protection fixed gcc version (CVE-2023-4039)
+
 * Thu Sep 07 2023 Betty Lakes <bettylakes@microsoft.com> - 9.16.37-1
 - Bump the version to 9.16.37 to fix CVE-2022-3924, CVE-2022-3094, CVE-2022-3736
